Remove unused recording service code
I initially thought the node with the microphone would be the server
and its clients would stream recordings down
for transcription, etc.

Then I changed my mind and inverted the flow;
now the node with the microphone is the hub and makes RPC requests for transcription.

The audio streaming code is therefore currently not necessary.