Parallelize recording pipeline
Previously, all recording steps in the pipeline were handled in sequence,
and processing the next recording didn't start until the previous
recording was completely handled.

WIth this commit, a new coroutine is launched as soon as a new recording
comes in,
and it handles its recording independently (and in parallel) from others.