Ensure external command output is properly handled
The way I was handling the external command's output was a potential
problem, in which it was possible for the external command to never exit.
See, for example: https://stackoverflow.com/q/9885643

The root cause was that I was calling Process.waitFor before collecting the output.
However, the output buffer was filling up, at which point there was a deadlock:
waitFor would not exit until the buffer were cleared,
but I didn't read from the buffer until waitFor was done.

The solution was to read from the buffer before calling waitFor.
This does have the side effect of waitFor not actually enforcing a limit
for how long the external program runs
(unless it happens to close stdout).
For now, that's not a major concern, so I'm not addressing it.